refactor: [US-044b] - Extract enrichment/display helpers into core/views/_helpers.py
